考试模块

+++ src/main/java/com/swpu/constructionsitesafety/controller/UserController.java
@@ -2,7 +2,10 @@ package com.swpu.constructionsitesafety.controller;
 
 import com.swpu.constructionsitesafety.context.BaseContext;
 import com.swpu.constructionsitesafety.entity.User;
-import com.swpu.constructionsitesafety.entity.dto.*;
+import com.swpu.constructionsitesafety.entity.dto.LoginDTO;
+import com.swpu.constructionsitesafety.entity.dto.UpdatePasswordDTO;
+import com.swpu.constructionsitesafety.entity.dto.UpdatePhoneDTO;
+import com.swpu.constructionsitesafety.entity.dto.UserIdDTO;
 import com.swpu.constructionsitesafety.entity.vo.LoginVO;
 import com.swpu.constructionsitesafety.service.IUserService;
 import com.swpu.constructionsitesafety.utils.ResultData;
@@ -75,8 +78,12 @@ public class UserController {
 
 	@PostMapping("/deleteUser")
 	public ResultData<Boolean> deleteUser(@RequestBody UserIdDTO userIdDTO) {
-		boolean result = userService.removeById(userIdDTO.getUserId());
-		return ResultData.success(result);
+		User user = userService.getById(BaseContext.getUserId());
+		if (user.getAuthority() == 1) {
+			boolean result = userService.removeById(userIdDTO.getUserId());
+			return ResultData.success(result);
+		}
+		return ResultData.fail(RC403.getCode(), RC403.getMessage());
 	}
 
 	@GetMapping("/getInfo")
